重构 显然需要修改构建脚本
