--- main.spec
@@ -1,45 +1,0 @@
-# -*- mode: python ; coding: utf-8 -*-
-import platform
-
-
-a = Analysis(
-    ['main.py'],
-    pathex=[],
-    binaries=[],
-    datas=[],
-    hiddenimports=[],
-    hookspath=[],
-    hooksconfig={},
-    runtime_hooks=[],
-    excludes=[],
-    noarchive=False,
-    optimize=2,
-)
-pyz = PYZ(a.pure)
-
-if platform.system() != 'Windows':
-    strip = True
-    a.exclude_system_libraries()
-else:
-    strip = False
-
-exe = EXE(
-    pyz,
-    a.scripts,
-    a.binaries,
-    a.datas,
-    [('O', None, 'OPTION'), ('O', None, 'OPTION')],
-    name='main',
-    debug=False,
-    bootloader_ignore_signals=False,
-    strip=strip,
-    upx=True,
-    upx_exclude=[],
-    runtime_tmpdir=None,
-    console=True,
-    disable_windowed_traceback=False,
-    argv_emulation=False,
-    target_arch=None,
-    codesign_identity=None,
-    entitlements_file=None,
-)
